+++ examples/crypto_ccs/protocols/enc_and_hmac/enc_and_hmac.c
@@ -45,7 +45,7 @@ void sender(char *enc_key, char *hmac_key, char *msg, unsigned int msg_len)
 
   {
     int message_len = 16 + (int) msg_len + 64;
-    char* message = malloc(message_len);
+    char* message = malloc((size_t)message_len);
     if (message == 0) abort();
 
     // IV stuff
@@ -154,7 +154,7 @@ int receiver(char *enc_key, char *hmac_key, char *msg)
 
   {
     int max_size = 16 + MAX_SIZE + 64;
-    char *buffer = malloc (max_size); if (buffer == 0) abort();
+    char *buffer = malloc ((size_t)max_size); if (buffer == 0) abort();
     size = net_recv(&socket2, buffer, (unsigned int) max_size);
     if (size <= 16 + 64) abort();
     enc_size = size - 16 - 64;
